# swagger-latest.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ParksComputing Content API",
    "description": "RESTful hypermedia API for managing site content (markdown pages/posts).",
    "version": "v1"
  },
  "paths": {
...
    "/api/content": {
      "get": {
        "tags": [
          "Content"
        ],
        "summary": "List content items (paged).",
        "parameters": [
          {
            "name": "prefix",
            "in": "query",
            "description": "Optional path/slug prefix filter.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "1-based page index (>=1).",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Items per page (1-100).",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "ETag": {
                "description": "Entity or aggregate ETag",
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "description": "Cache directives",
                "schema": {
                  "type": "string"
                }
              },
              "X-Total-Count": {
                "description": "Total items for current filter",
                "schema": {
                  "type": "string"
                }
              },
              "X-Page": {
                "description": "Current page",
                "schema": {
                  "type": "string"
                }
              },
              "X-Page-Size": {
                "description": "Page size",
                "schema": {
                  "type": "string"
                }
              },
              "X-RateLimit-Limit": {
                "description": "Request limit per window (illustrative)",
                "schema": {
                  "type": "string"
                }
              },
              "X-RateLimit-Remaining": {
                "description": "Remaining requests in window",
                "schema": {
                  "type": "string"
                }
              },
              "X-RateLimit-Reset": {
                "description": "Epoch seconds window resets",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/xfer": {
                "schema": {
                  "type": "string",
                  "description": "XferLang representation",
                  "example": "Microsoft.OpenApi.Any.OpenApiString"
                },
                "example": "Microsoft.OpenApi.Any.OpenApiString",
                "examples": {
                  "xfer": {
                    "summary": "XferLang",
                    "description": "XferLang media representation",
                    "value": "(\r\n  {\r\n    id \"sample-slug\"\r\n    slug \"string\"\r\n    title \"string\"\r\n    description <\"Sample description\">\r\n    createdUtc \"string\"\r\n    updatedUtc \"string\"\r\n    language \"string\"\r\n    rawMarkdown \"string\"\r\n    rawHtml \"string\"\r\n    published ~true \r\n    links (\r\n      {\r\n        rel \"string\"\r\n        method \"string\"\r\n        href \"string\"\r\n      }\r\n    )\r\n  }\r\n)"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContentResource"
                  }
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests"
          }
        }
      },
      "post": {
        "tags": [
          "Content"
        ],
        "requestBody": {
          "content": {
            "application/xfer": {
              "schema": {
                "type": "string",
                "description": "XferLang representation",
                "example": "Microsoft.OpenApi.Any.OpenApiString"
              },
              "example": "Microsoft.OpenApi.Any.OpenApiString",
              "examples": {
                "xfer": {
                  "summary": "XferLang",
                  "description": "XferLang media representation",
                  "value": "{\r\n  slug \"string\"\r\n  title \"string\"\r\n  description <\"Sample description\">\r\n  language \"string\"\r\n  bodyMarkdown \"string\"\r\n}"
                }
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/xfer": {
                "schema": {
                  "type": "string",
                  "description": "XferLang representation",
                  "example": "Microsoft.OpenApi.Any.OpenApiString"
                },
                "example": "Microsoft.OpenApi.Any.OpenApiString",
                "examples": {
                  "xfer": {
                    "summary": "XferLang",
                    "description": "XferLang media representation",
                    "value": "{\r\n  id \"sample-slug\"\r\n  slug \"string\"\r\n  title \"string\"\r\n  description <\"Sample description\">\r\n  createdUtc \"string\"\r\n  updatedUtc \"string\"\r\n  language \"string\"\r\n  rawMarkdown \"string\"\r\n  rawHtml \"string\"\r\n  published ~true \r\n  links (\r\n    {\r\n      rel \"string\"\r\n      method \"string\"\r\n      href \"string\"\r\n    }\r\n  )\r\n}"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentResource"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/xfer": {
                "schema": {
                  "type": "string",
                  "description": "XferLang representation",
                  "example": "Microsoft.OpenApi.Any.OpenApiString"
                },
                "example": "Microsoft.OpenApi.Any.OpenApiString"
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://httpstatuses.com/409",
                  "title": "Conflict",
                  "status": 409,
                  "detail": "Resource already exists",
                  "instance": "/api/content/{id}"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://httpstatuses.com/401",
                  "title": "Unauthorized",
                  "status": 401,
                  "detail": "Authentication required or invalid token",
                  "instance": "/api/content/{id}"
                }
              },
              "application/xfer": {
                "schema": {
                  "type": "string",
                  "description": "XferLang representation",
                  "example": "Microsoft.OpenApi.Any.OpenApiString"
                },
                "example": "Microsoft.OpenApi.Any.OpenApiString"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://httpstatuses.com/403",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "Insufficient permissions",
                  "instance": "/api/content/{id}"
                }
              },
              "application/xfer": {
                "schema": {
                  "type": "string",
                  "description": "XferLang representation",
                  "example": "Microsoft.OpenApi.Any.OpenApiString"
                },
                "example": "Microsoft.OpenApi.Any.OpenApiString"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://httpstatuses.com/404",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "Resource not found",
                  "instance": "/api/content/{id}"
                }
              },
              "application/xfer": {
                "schema": {
                  "type": "string",
                  "description": "XferLang representation",
                  "example": "Microsoft.OpenApi.Any.OpenApiString"
                },
                "example": "Microsoft.OpenApi.Any.OpenApiString"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ContentResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "slug": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "createdUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "language": {
            "type": "string",
            "nullable": true
          },
          "rawMarkdown": {
            "type": "string",
            "nullable": true
          },
          "rawHtml": {
            "type": "string",
            "nullable": true
          },
          "published": {
            "type": "boolean"
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LinkStub"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateContentRequest": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "language": {
            "type": "string",
            "nullable": true
          },
          "bodyMarkdown": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LinkStub": {
        "type": "object",
        "properties": {
          "rel": {
            "type": "string",
            "nullable": true
          },
          "method": {
            "type": "string",
            "nullable": true
          },
          "href": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter 'Bearer {token}'",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}